requirements.txt
certifi==2019.9.11
patool==1.12
pipenv==2018.11.26
virtualenv==16.7.7
virtualenv-clone==0.5.3
